(refactor) Move #getCnnFeed to its own file.
